backend/requirements.txt
# Core web framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.6,<1.0.0

# HTTP client and utilities
httpx>=0.25.0,<1.0.0
aiofiles>=23.0.0,<25.0.0
requests>=2.28.0,<3.0.0

# Data handling
PyYAML>=6.0.0,<7.0.0
python-dateutil>=2.8.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0

# LLM Providers
openai>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0
anthropic>=0.7.0,<1.0.0

# Multi-agent orchestration
langgraph>=0.0.60,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-openai>=0.0.8,<1.0.0

# Document processing
pypdf>=3.15.0,<5.0.0
python-docx>=1.1.0,<2.0.0
python-pptx>=0.6.21,<1.0.0
pandas>=1.5.0,<3.0.0
Pillow>=9.0.0,<12.0.0
pdf2image>=1.16.0,<4.0.0
PyMuPDF>=1.23.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# Database (sqlite3 is built-in with Python)
# No additional database dependencies required

# Utilities
numpy>=1.21.0,<2.0.0

# Development and testing
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<25.0.0
flake8>=6.0.0,<8.0.0
mypy>=1.5.0,<2.0.0

# Production deployment
gunicorn>=20.1.0,<23.0.0

# Additional dependencies for Python 3.12 compatibility
typing-extensions>=4.5.0,<5.0.0